Layers: Implement proxying of show/simplified checkboxes, so rnet/rnet-simplified can be isolated
Previously there was logic which handles special toggling of normal/simplified modes.
This commit changes this so that the two variants have a normal on/off checkbox, each of which is set by the normal/simplified checkboxes.
This will unlock generic handling of layer visibility, as the special toggling mode was different to other layers.